requirements-dev.txt
django==3.1.6
stripe==2.56.0
whitenoise==5.2.0
django-sslserver==0.22
python-dotenv==0.19.0

black==19.3b0
flake8==3.9.0
flake8-bugbear==20.11.1
isort==5.7.0
pylint==2.6.0
pylint-django==2.4.2
